+++ com_joomlaquiz/admin/views/questions/view.html.php
@@ -137,20 +137,30 @@ class JoomlaquizViewQuestions extends JViewLegacy
     }
 	
 	protected function addCopyToolBar(){
-		JToolBarHelper::cancel('question.cancel', 'JTOOLBAR_CANCEL');
-		JToolBarHelper::custom('questions.copy_question', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', false);
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+        JToolBarHelper::cancel('question.cancel', 'JTOOLBAR_CANCEL');
+        if ($canDo->get('core.create') && $canDo->get('core.edit')) {
+            JToolBarHelper::custom('questions.copy_question', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', false);
+        }
 	}
 	
 	protected function addMoveToolBar()
 	{
-		JToolBarHelper::cancel('question.cancel', 'JTOOLBAR_CANCEL');
-		JToolBarHelper::custom('questions.move_question', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', false);
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+        JToolBarHelper::cancel('question.cancel', 'JTOOLBAR_CANCEL');
+        if ($canDo->get('core.create') && $canDo->get('core.edit')) {
+            JToolBarHelper::custom('questions.move_question', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', false);
+        }
 	}
 	
 	protected function addUploadquestToolBar()
 	{
-		JToolBarHelper::cancel('questions.cancel', 'JTOOLBAR_CANCEL');
-		JToolBarHelper::custom('questions.uploadquestions', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_UPLOAD', false);
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+
+        JToolBarHelper::cancel('questions.cancel', 'JTOOLBAR_CANCEL');
+        if ($canDo->get('core.create') && $canDo->get('core.edit')) {
+            JToolBarHelper::custom('questions.uploadquestions', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_UPLOAD', false);
+        }
 	}
 	
     /**
@@ -158,20 +168,29 @@ class JoomlaquizViewQuestions extends JViewLegacy
     */
     protected function addToolBar() 
     {
-        $bar = JToolBar::getInstance('toolbar'); 
-		$bar->appendButton( 'Custom', '<div id="toolbar-new" class="btn-group"><a class="btn btn-small btn-success" onclick="javascript: tb_start(this);return false;" href="index.php?option=com_joomlaquiz&amp;tmpl=component&amp;task=questions.new_question_type&amp;KeepThis=true&amp;TB_iframe=true&amp;height=350&amp;width=700" href="#"><i class="icon-new icon-white"></i>'.JText::_('COM_JOOMLAQUIZ_NEW').'</a></div>');
-        JToolBarHelper::editList('question.edit');
-        JToolBarHelper::divider();
-		JToolBarHelper::custom('questions.publish', 'publish.png', 'publish_f2.png','JTOOLBAR_PUBLISH', true);
-		JToolBarHelper::custom('questions.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
-        JToolBarHelper::divider();
-        JToolBarHelper::deleteList('', 'questions.checkComplitedQuestions');
-		JToolBarHelper::divider();
-		JToolBarHelper::custom('questions.move_question_sel', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', true);
-		JToolBarHelper::custom('questions.copy_question_sel', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', true);
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+        $bar = JToolBar::getInstance('toolbar');
+        if ($canDo->get('core.create')) {
+            $bar->appendButton('Custom',
+                '<div id="toolbar-new" class="btn-group"><a class="btn btn-small btn-success" onclick="javascript: tb_start(this);return false;" href="index.php?option=com_joomlaquiz&amp;tmpl=component&amp;task=questions.new_question_type&amp;KeepThis=true&amp;TB_iframe=true&amp;height=350&amp;width=700" href="#"><i class="icon-new icon-white"></i>' . JText::_('COM_JOOMLAQUIZ_NEW') . '</a></div>');
+        }
+        if ($canDo->get('core.edit')) {
+            JToolBarHelper::editList('question.edit');
+            JToolBarHelper::custom('questions.publish', 'publish.png', 'publish_f2.png', 'JTOOLBAR_PUBLISH', true);
+            JToolBarHelper::custom('questions.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
+        }
+        if ($canDo->get('core.delete')) {
+            JToolBarHelper::deleteList('', 'questions.checkComplitedQuestions');
+        }
+        if ($canDo->get('core.create') && $canDo->get('core.edit')) {
+            JToolBarHelper::custom('questions.move_question_sel', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', true);
+            JToolBarHelper::custom('questions.copy_question_sel', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', true);
+        }
 		JToolBarHelper::custom('questions.quizzes', 'previous.png', 'previous_f2.png', 'COM_JOOMLAQUIZ_QUIZZES', false);
-		JToolBarHelper::custom('questions.hotspot_converter', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_HOTSPOT_CONVERTER', true);
-        $bar->appendButton( 'Custom', '<div class="btn-group"><a class="btn btn-small" onclick="javascript: tb_start(this);return false;" href="index.php?option=com_joomlaquiz&amp;tmpl=component&amp;view=configuration&amp;KeepThis=true&amp;TB_iframe=true&amp;height=350&amp;width=700" href="#"><i class="icon-options"></i>'.JText::_('COM_JOOMLAQUIZ_OPTIONS').'</a></div>');
+        if ($canDo->get('core.create')) {
+            JToolBarHelper::custom('questions.hotspot_converter', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_HOTSPOT_CONVERTER', true);
+            $bar->appendButton('Custom', '<div class="btn-group"><a class="btn btn-small" onclick="javascript: tb_start(this);return false;" href="index.php?option=com_joomlaquiz&amp;tmpl=component&amp;view=configuration&amp;KeepThis=true&amp;TB_iframe=true&amp;height=350&amp;width=700" href="#"><i class="icon-options"></i>' . JText::_('COM_JOOMLAQUIZ_OPTIONS') . '</a></div>');
+        }
     }
 	
 	protected function getSortFields()
