+++ com_joomlaquiz/admin/views/quizzes/view.html.php
@@ -93,38 +93,61 @@ class JoomlaquizViewQuizzes extends JViewLegacy
     }
 	
 	protected function addCopyToolBar(){
-		JToolBarHelper::cancel('quiz.cancel', 'JTOOLBAR_CANCEL');
-		JToolBarHelper::custom('quizzes.copy_quizzes', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', false);
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+        JToolBarHelper::cancel('quiz.cancel', 'JTOOLBAR_CANCEL');
+        if ($canDo->get('core.create') && $canDo->get('core.edit')) {
+            JToolBarHelper::custom('quizzes.copy_quizzes', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', false);
+        }
 	}
 	
 	protected function addMoveToolBar() {
-		JToolBarHelper::cancel('quiz.cancel', 'JTOOLBAR_CANCEL');
-		JToolBarHelper::custom('quizzes.move_quizzes', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', false);
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+        JToolBarHelper::cancel('quiz.cancel', 'JTOOLBAR_CANCEL');
+        if ($canDo->get('core.create') && $canDo->get('core.edit')) {
+            JToolBarHelper::custom('quizzes.move_quizzes', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', false);
+        }
 	}
 		
-	protected function addImportToolBar(){
-		JToolBarHelper::cancel('quiz.cancel', 'JTOOLBAR_CANCEL');
-		JToolBarHelper::custom('quizzes.import_quizzes', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_IMPORT', false);
-	}
+	protected function addImportToolBar()
+    {
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+        JToolBarHelper::cancel('quiz.cancel', 'JTOOLBAR_CANCEL');
+        if ($canDo->get('core.create')) {
+            JToolBarHelper::custom('quizzes.import_quizzes', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_IMPORT',
+                false);
+        }
+    }
 	
     /**
     * Setting the toolbar
     */
     protected function addToolBar() 
     {
-        JToolBarHelper::addNew('quiz.add');
-        JToolBarHelper::editList('quiz.edit');
-        JToolBarHelper::divider();
-		JToolBarHelper::custom('quizzes.publish', 'publish.png', 'publish_f2.png','JTOOLBAR_PUBLISH', true);
-		JToolBarHelper::custom('quizzes.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
-        JToolBarHelper::divider();
-        JToolBarHelper::deleteList('', 'quizzes.delete');
-		JToolBarHelper::divider();
-		JToolBarHelper::custom('quizzes.export_quizzes', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_EXPORT', true);
-		JToolBarHelper::custom('quizzes.export_quizzes_all', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_EXPORT_ALL', false);
-		JToolBarHelper::custom('quizzes.move_quiz_sel', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', true);
-		JToolBarHelper::custom('quizzes.copy_quiz_sel', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', true);
-		JToolBarHelper::custom('quizzes.quizcategories', 'previous.png', 'previous_f2.png', 'COM_JOOMLAQUIZ_CATEGORIES', false);
+        $canDo = JHelperContent::getActions('com_joomlaquiz', 'component');
+
+        if ($canDo->get('core.create'))
+        {
+            JToolBarHelper::addNew('quiz.add');
+        }
+        if ($canDo->get('core.edit')) {
+            JToolBarHelper::editList('quiz.edit');
+            JToolBarHelper::custom('quizzes.publish', 'publish.png', 'publish_f2.png', 'JTOOLBAR_PUBLISH', true);
+            JToolBarHelper::custom('quizzes.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
+        }
+        if ($canDo->get('core.delete')) {
+            JToolBarHelper::deleteList('', 'quizzes.delete');
+        }
+        if ($canDo->get('core.edit')) {
+            JToolBarHelper::custom('quizzes.export_quizzes', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_EXPORT', true);
+            JToolBarHelper::custom('quizzes.export_quizzes_all', 'featured.png', 'featured_f2.png', 'COM_JOOMLAQUIZ_EXPORT_ALL', false);
+        }
+        if ($canDo->get('core.create') && $canDo->get('core.edit')) {
+            JToolBarHelper::custom('quizzes.move_quiz_sel', 'move.png', 'move_f2.png', 'COM_JOOMLAQUIZ_MOVE', true);
+            JToolBarHelper::custom('quizzes.copy_quiz_sel', 'copy.png', 'copy_f2.png', 'COM_JOOMLAQUIZ_COPY', true);
+        }
+
+        JToolBarHelper::custom('quizzes.quizcategories', 'previous.png', 'previous_f2.png', 'COM_JOOMLAQUIZ_CATEGORIES', false);
+
     }
 	
 	protected function getSortFields()
